+++ resources/lang/en/permissions.php
@@ -3,6 +3,7 @@
 return [
     'exception' => [
         'delete' => 'Delete',
+        'update' => 'Update',
         'view' => 'View',
     ],
 ];