+++ resources/lang/ru/permissions.php
@@ -3,6 +3,7 @@
 return [
     'exception' => [
         'delete' => 'Удаление',
+        'update' => 'Изменение',
         'view' => 'Просмотр',
     ],
 ];